+++ github-runner-manager/src/github_runner_manager/manager/runner_scaler.py
@@ -114,7 +114,7 @@ class RunnerScaler:
         Returns:
             A new RunnerScaler.
         """
-        labels = application_configuration.extra_labels
+        labels = list(application_configuration.extra_labels)
         server_config = None
         base_quantity = 0
         if combinations := application_configuration.non_reactive_configuration.combinations: