+++ docs/changelog.md
@@ -1,9 +1,16 @@
 # Changelog
 
+
+### 2025-05-27
+
+- The charm will reconcile the state of the self-hosted runners in the background when it is able to.
+- Removed the reconcile-runners juju action.
+
 ### 2025-05-22
 
 - Add possibility to run a script in the pre-job phase of a runner. This can be useful to setup 
 network/infrastructure specific things.
+
 
 ### 2025-05-09